third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/HTMLStyleElement/style-onload-remove-crash-expected.txt

CONSOLE ERROR: line 15: Uncaught TypeError: myDiv.createShadowRoot is not a function
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
